+++ src/elements/Product.php
@@ -301,9 +301,19 @@ class Product extends Element
     protected function isEditable(): bool
     {
         if ($this->getType()) {
-            $uid = $this->getType()->uid;
+            return Craft::$app->getUser()->checkPermission('commerce-manageProductType:' . $this->getType()->uid);
+        }
 
-            return Craft::$app->getUser()->checkPermission('commerce-manageProductType:' . $uid);
+        return false;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function isDeletable(): bool
+    {
+        if ($this->getType()) {
+            return Craft::$app->getUser()->checkPermission('commerce-manageProductType:' . $this->getType()->uid); // TODO: Change this to a real delete permission in 4.0
         }
 
         return false;